feat: adding repositories and controllers
