+++ src/main/resources/application.yml
@@ -1,9 +1,22 @@
-spring.datasource.url=jdbc:h2:mem:testdb
-spring.datasource.driverClassName=org.h2.Driver
-spring.datasource.username=admin
-spring.datasource.password=admin
-spring.h2.console.enabled=true
-spring.datasource.initialization-mode=always
-spring.h2.console.path=/h2-console
-spring.h2.console.settings.web-allow-others=true
-server.port=8080
+spring:
+  datasource:
+    url: jdbc:h2:mem:testdb
+    driverClassName: org.h2.Driver
+    username: admin
+    password: admin
+    initialization-mode: always
+
+  h2:
+    console:
+      enabled: true
+      path: /h2-console
+      settings:
+        web-allow-others: true
+  jpa:
+    show-sql: true
+
+server:
+  port: 8080
+
+via_cep:
+  url: https://viacep.com.br/ws
